+++ src/lib/djerba/helpers/provenance_helper/helper.py
@@ -190,6 +190,7 @@ class main(helper_base):
                 self.WG_N_IDX: reader.parse_wg_index_ref_path()
             },
             reader.WF_DELLY: reader.parse_delly_path(),
+            reader.WF_HRDETECT: reader.parse_hrdetect_path(),
             reader.WF_MAVIS: reader.parse_mavis_path(),
             reader.WF_MRDETECT: reader.parse_mrdetect_path(),
             reader.WF_MSISENSOR: reader.parse_msi_path(),